New CQC & movement system
This change includes quite a few notable changes.
If AI engage in combat within 150m of each other, they will from there on have two choices on how to proceed. If they outnumber the enemy, they will push the advantage, and if they are outnumbered they will attempt to fall back.

Furthermore, vanilla Suppression is back and Autocombat is disabled. Autocombat will be enabled while in CQC, though outside of it it will be disabled in order to keep the AI moving quickly.

In the close future, there will be a new Suppression system that builds upon the Vanilla suppression system to make midrange and further combat more interesting.

To best utilize these changes, I recommend turning down AI "commanding" and "courage" down to 0, in order to make them more susceptible to suppressive fire.